Working Login with DB connection
