+++ src/HomePage.java
@@ -92,11 +92,6 @@ public class HomePage extends JFrame {
     }
 
     public static void main(String[] args) {
-        SwingUtilities.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-                new HomePage();
-            }
-        });
+        SwingUtilities.invokeLater(HomePage::new);
     }
 }
