--- src/SystemAdminLoginUI.java
@@ -1,99 +1,0 @@
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-public class SystemAdminLoginUI extends JFrame {
-    // Dummy credentials for demonstration purposes
-    private static final String ADMIN_USERNAME = "admin";
-    private static final String ADMIN_PASSWORD = "password123";
-
-    private JTextField usernameField;
-    private JPasswordField passwordField;
-
-    public SystemAdminLoginUI() {
-        setTitle("System Admin Login");
-        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setSize(900, 600);
-        setLocationRelativeTo(null); // Center the window
-
-        // Create components
-        JPanel panel = new JPanel();
-        panel.setLayout(new GridBagLayout());
-
-        GridBagConstraints constraints = new GridBagConstraints();
-        constraints.insets = new Insets(10, 10, 10, 10);
-
-        JLabel usernameLabel = new JLabel("Username:");
-        usernameField = new JTextField(20);
-        JLabel passwordLabel = new JLabel("Password:");
-        passwordField = new JPasswordField(20);
-        JButton loginButton = new JButton("Login");
-
-        // Add components to the panel with constraints
-        constraints.gridx = 0;
-        constraints.gridy = 0;
-        panel.add(usernameLabel, constraints);
-
-        constraints.gridx = 1;
-        constraints.gridy = 0;
-        panel.add(usernameField, constraints);
-
-        constraints.gridx = 0;
-        constraints.gridy = 1;
-        panel.add(passwordLabel, constraints);
-
-        constraints.gridx = 1;
-        constraints.gridy = 1;
-        panel.add(passwordField, constraints);
-
-        constraints.gridx = 0;
-        constraints.gridy = 2;
-        constraints.gridwidth = 2;
-        constraints.anchor = GridBagConstraints.CENTER;
-        panel.add(loginButton, constraints);
-
-        // Add action listener to the login button
-        loginButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                login();
-            }
-        });
-
-        // Add panel to the frame
-        add(panel);
-
-        setVisible(true);
-    }
-
-    private void login() {
-        String username = usernameField.getText();
-        String password = new String(passwordField.getPassword());
-
-        // Check if the entered credentials match the admin credentials
-        if (username.equals(ADMIN_USERNAME) && password.equals(ADMIN_PASSWORD)) {
-            JOptionPane.showMessageDialog(this, "Login successful! You have accessed administrative functions.");
-            // Call the method to access administrative functions
-            accessAdminFunctions();
-        } else {
-            JOptionPane.showMessageDialog(this, "Invalid username or password. Please try again.", "Error", JOptionPane.ERROR_MESSAGE);
-        }
-    }
-
-    // Method to simulate accessing administrative functions
-    private void accessAdminFunctions() {
-        // Add your administrative functionalities here
-        // For demonstration, let's just print a message
-        System.out.println("Administrative functions accessed successfully.");
-    }
-
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-                new SystemAdminLoginUI();
-            }
-        });
-    }
-}
